testbed.MeasureTest now displays the adjacency matrix and cost graph of each measure being tested. Fixed a bug on line 13 of MiddleburyData.BodyReference that was causing some trouble in testing.

+++ components/+MiddleburyData/BodyReference.m
@@ -10,7 +10,7 @@ classdef BodyReference < MiddleburyData.MiddleburyDataConfig & tom.Trajectory
 
   methods (Access = public, Static = true)
     function this = BodyReference(initialTime)
-      this.interval = tom.TimeInterval(initialTime, tom.WorldTime(initialTime+(this.numImages-1)/this.fps));
+      this.interval = tom.TimeInterval(initialTime, tom.WorldTime(initialTime+(double(this.numImages)-1)/this.fps));
     end
   end
